added the Transaction immutable class
The Investment.java class now returns an immutable object of type Transaction during a purchase/sell. Also it raises an exception if number of shares you want to buy is greater than how many you are able to buy, or if the number of shares you want to sell is greater than how many you own.